requirements.txt
alembic==0.9.6
bitcoin==1.1.42
click==6.7
decorator==4.1.2
Flask==0.12.2
Flask-Admin==1.5.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.0
Flask-Migrate==2.1.1
Flask-QRcode==1.0.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
olefile==0.44
pbr==3.1.1
Pillow==4.3.0
pycoin==0.80
python-dateutil==2.4.2
python-editor==1.0.3
python-bittrex==0.3.0
qrcode==5.3
requests==2.9.1
six==1.11.0
SQLAlchemy==1.1.15
sqlalchemy-migrate==0.11.0
sqlparse==0.2.4
Tempita==0.5.2
Werkzeug==0.12.2
WTForms==2.1
